backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.5,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "biopython (>=1.85,<2.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "psycopg-c (>=3.2.6,<4.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.server:poetry_run_dev"
